Move `EvaluateTemplate` to a separate file
This patch moves `EvaluateTemplate` function used for testing from the
`servers.go` file to a new `templates.go` file. This new file will be
used to contain other functions related to testing with templates.
